refactor: extract commands to new file
