+++ client/src/extension.ts
@@ -7,7 +7,6 @@ import {
   StatusBarAlignment,
   ThemeColor,
   StatusBarItem,
-  Uri,
 } from "vscode";
 
 import {
@@ -15,10 +14,9 @@ import {
   LanguageClientOptions,
   ServerOptions,
   TransportKind,
-  ExecuteCommandRequest,
 } from "vscode-languageclient/node";
 
-import { objectToSnake } from "ts-case-convert";
+import { format, syncSettings } from "./command";
 
 let client: LanguageClient;
 
@@ -65,84 +63,16 @@ export function activate(context: ExtensionContext) {
   );
 
   context.subscriptions.push(
-    commands.registerCommand("uroborosql-fmt.uroborosql-format", async () => {
-      const uri = window.activeTextEditor.document.uri;
-      const version = window.activeTextEditor.document.version;
-      const selections = window.activeTextEditor.selections;
-
-      await client.sendRequest(ExecuteCommandRequest.type, {
-        command: "uroborosql-fmt.executeFormat",
-        arguments: [uri, version, selections],
-      });
-    }),
+    commands.registerCommand(
+      "uroborosql-fmt.uroborosql-format",
+      format(client),
+    ),
   );
-
-  type OptionsRecord = Record<string, boolean | number | string>;
-
-  const isFileExists = async (uri: Uri): Promise<boolean> => {
-    try {
-      await workspace.fs.stat(uri);
-      return true;
-    } catch (_) {
-      return false;
-    }
-  };
 
   context.subscriptions.push(
     commands.registerCommand(
       "uroborosql-fmt.sync-config-file-with-extension-config",
-      async () => {
-        // uroborosql-fmt の設定を取得
-        const vsCodeConfig = workspace.getConfiguration("uroborosql-fmt");
-
-        // Default value of `uroborosql-fmt.configurationFilePath` is "".
-        const vsCodeConfigPath: string = vsCodeConfig.get(
-          "configurationFilePath",
-        );
-        const configFilePath =
-          vsCodeConfigPath !== "" ? vsCodeConfigPath : ".uroborosqlfmtrc.json";
-
-        // VSCodeで開いているディレクトリを取得
-        // 開いていない場合はエラーを出して終了
-        const folders = workspace.workspaceFolders;
-        if (folders === undefined) {
-          window.showErrorMessage(
-            "Error: Open the folder before executing this command.",
-          );
-          return;
-        }
-        const folderPath = folders[0].uri;
-        const configFileFullPath = Uri.joinPath(folderPath, configFilePath);
-
-        // 設定値 `configurationFilePath` の除外・ value が null のエントリを除外・ WorkSpaceConfiguration が持つメソッドと内部オブジェクトを除外
-        const formattingConfig: OptionsRecord = Object.fromEntries(
-          Object.entries(vsCodeConfig).filter(
-            ([key, value]) =>
-              key !== "configurationFilePath" &&
-              value !== null &&
-              typeof value !== "function" &&
-              typeof value !== "object",
-          ),
-        );
-
-        let existingConfig: OptionsRecord;
-        if (await isFileExists(configFileFullPath)) {
-          const file = await workspace.fs.readFile(configFileFullPath);
-          existingConfig = JSON.parse(file.toString());
-        } else {
-          existingConfig = {};
-        }
-
-        // 明示的に設定したものだけを上書きする
-        const merged = {
-          ...existingConfig,
-          ...objectToSnake(formattingConfig),
-        };
-        const content = JSON.stringify(merged, null, 2);
-
-        const blob: Uint8Array = Buffer.from(content);
-        await workspace.fs.writeFile(configFileFullPath, blob);
-      },
+      syncSettings,
     ),
   );
